refactor: remove deprecated orders flow and improve handler naming
  Remove the orders topic integration (CoreAdapter, OrderMessageHandler,
  CreateInvoiceCommand) that was replaced by the sales.confirmed event flow.
  Rename handlers and commands for clarity: InvoiceMessageHandler →
  InvoiceEventHandler, SaleConfirmedMessageHandler → SaleConfirmedHandler,
  CreateInvoiceFromSaleCommand → CreateInvoiceCommand. Clean up unused
  shared code (BaseHttpClient, Endpoint type) and core config references.
