+++ src/modules/invoice/app/types.ts
@@ -1,9 +1,3 @@
-export interface CreateInvoiceInput {
-  id: number;
-  accessCode: string;
-  sequence: string;
-}
-
 export interface InvoiceMessageInput {
   type: string;
   invoiceId: string;
