+++ src/modules/invoice/infra/messaging/topics.ts
@@ -1,5 +1,4 @@
 export enum KAFKA_TOPICS {
-  ORDERS = 'orders',
   INVOICES = 'invoices',
   SALES_CONFIRMED = 'sales.confirmed',
 }
