--- src/shared/types.ts
@@ -1,4 +1,0 @@
-export type Endpoint = {
-  host: string;
-  timeout: number;
-};
